Fix category test in main activity
