+++ app/src/androidTest/java/com/hitherejoe/vineyard/util/CustomMatchers.java
@@ -2,16 +2,39 @@ package com.hitherejoe.vineyard.util;
 
 import android.os.IBinder;
 import android.support.test.espresso.Root;
+import android.text.TextUtils;
+import android.view.View;
 import android.view.WindowManager;
+
+import com.hitherejoe.vineyard.R;
 
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 
-/**
- * Created by hitherejoe on 25/10/15.
- */
+import static android.support.test.espresso.core.deps.guava.base.Preconditions.checkArgument;
+import static android.support.test.espresso.matcher.ViewMatchers.isDescendantOfA;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static org.hamcrest.Matchers.allOf;
+
 public class CustomMatchers {
+
+    public static Matcher<View> withItemText(final String itemText, final int parentId) {
+        checkArgument(!TextUtils.isEmpty(itemText), "itemText cannot be null or empty");
+        return new TypeSafeMatcher<View>() {
+            @Override
+            public boolean matchesSafely(View item) {
+                return allOf(isDescendantOfA(withId(parentId)),
+                        withText(itemText)).matches(item);
+            }
+
+            @Override
+            public void describeTo(Description description) {
+                description.appendText("is isDescendantOfA RecyclerView with text " + itemText);
+            }
+        };
+    }
 
     /**
      * Matcher that is Toast window.
